Added a rule for Lambda runtime deprecation.
AWS doesn't have a programmatic interface for runtime deprecation dates.
Consequently, I used a horrible kludge: hard-coding dates copied from
https://docs.aws.amazon.com/lambda/latest/dg/lambda-runtimes.html.  If
the table hasn't been updated in the last 180 days, ScoutSuite will warn
the user.

Owing to this kludge, this rule has *not* been added to the default
ruleset. However, it is in the detailed ruleset.  If AWS ever adds a
programmatic interface, we can update this rule and add it to the
default ruleset.