Unit test for the new AWS lambda deprecation rule.
This test handle only rule processing, not data acquisition (which is
where most of the logic is).

test_rules_processingengine.py only used rules from the default ruleset.
Since awslambda-runtime-deprecated is *not* in the default ruleset, I
repurposed some dead code relating to a test ruleset: the test ruleset
is now merged with the default ruleset when running this test.