Fixed the AWS rule "iam-no-support-role".
The existing rule did not work: the IAM facade only enumerated AWS-managed
permission policies that were attached to things so the rule would never
trigger.  Additionally, the existing rule allowed AWSSupportAccess to be
attached to a User or Group, whereas CIS and ADA specifically require it
to be attached to a Role.

Unfortunately, there is no way to call iam:ListPolicies for only a specific
permission policy; the only filtering is by attachment status (which
doesn't help) and by path (which also doesn't help since AWSSupportAccess
is at the root path).  So we have no choice but to pull all permission
policies, including the hundreds of AWS-managed policies that the account
doesn't use at all.  Then we end up pulling their contents and
attachments, which takes a long time.  I optimize this by skipping
unattached policies when querying for contents and attachments.  *That*
violated an assumption in the policy-parsing logic that all permission
policies are valid, so I patched that to skip empty policies (yielding
empty lists of statements, which seems to be handled safely everywhere
else).

With this done, I changed the rule's logic to fail if either
AWSSupportAccess is attached to something but not to a Role or if it is
not attached to anything at all.

I added the rule to the ADA-CP ruleset.  I dropped it from "danger" to
"warning" in the "detailed" ruleset because it's sort of a silly check as
it assumes that an account is being managed in a very specific way.